requirements.txt
alembic==1.13.1
aniso8601==9.0.1
anyio==4.4.0
asgiref==3.8.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
Flask==3.0.3
Flask-Log-Request-ID==0.10.1
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
packaging==24.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
typing_extensions==4.12.1
urllib3==2.2.1
uvicorn==0.30.1
Werkzeug==3.0.3
zipp==3.19.1
